requirements.txt
fastapi
uvicorn
pydantic
pandas
faiss-cpu
sentence-transformers
torch
numpy
scikit-learn